refactor: simplify PDP by merging mobile/desktop into responsive components
Collapse the dual mobile/desktop component trees into a single responsive
layout. MobileCarousel and ImageGrid become a unified ProductMedia component,
MobileBuyButtons is renamed to BuyButtons, and the size prop is replaced with
responsive Tailwind classes throughout.
